+++ src/main/java/eu/dissco/core/handlemanager/domain/requests/PostRequestData.java
@@ -3,7 +3,7 @@ package eu.dissco.core.handlemanager.domain.requests;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.databind.JsonNode;
-import eu.dissco.core.handlemanager.domain.requests.vocabulary.FdoType;
+import eu.dissco.core.handlemanager.domain.FdoType;
 
 public record PostRequestData(
     @JsonProperty(required = true)
