+++ src/test/java/eu/dissco/core/handlemanager/service/DataCiteServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.BDDMockito.then;
 
 import eu.dissco.core.handlemanager.domain.datacite.DataCiteEvent;
 import eu.dissco.core.handlemanager.domain.datacite.EventType;
-import eu.dissco.core.handlemanager.domain.requests.vocabulary.FdoType;
+import eu.dissco.core.handlemanager.domain.FdoType;
 import eu.dissco.core.handlemanager.properties.KafkaPublisherProperties;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
